105410 - Ajustes query mediator com handler duplicado e apontamento para mesmo request handler
